Fixed linter errors
