--- modules/storage.js
@@ -1,3 +1,0 @@
-export default function (data) {
-  localStorage.setItem('books', JSON.stringify(data));
-}
